Moved search helpers to separate file
